requirements.txt
torch>=2.2
transformers>=4.45
peft>=0.11
datasets>=2.20
accelerate>=0.30
bitsandbytes>=0.43
huggingface_hub>=0.24
nibabel>=5.2
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0
tqdm>=4.66
